+++ plugin.py
@@ -418,21 +418,7 @@ class PHPUnit():
                 self.view.settings().get('phpunit.text_ui_result_font_size')
             )
 
-        # BC: to be removed in v3.0.0
-        # Custom color schemes are deprecated and will
-        # be removed in v3.0.0. Instead, a definitive
-        # syntax for test results panels will be
-        # written and popular color schemes will
-        # be asked to support it. See the issue
-        # tracker for more details.
-        if self.view.settings().has('phpunit.color_scheme'):
-            color_scheme = self.view.settings().get('phpunit.color_scheme')
-            if color_scheme:
-                color_scheme = color_scheme.replace('Packages/phpunitkit/color-schemes/', 'Packages/phpunitkit/res/')
-            else:
-                color_scheme = self.view.settings().get('color_scheme')
-        else:
-            color_scheme = self.get_auto_generated_color_scheme()
+        color_scheme = self.get_auto_generated_color_scheme()
         self.window.create_output_panel('exec').settings().set('color_scheme', color_scheme)
 
     def run_last(self):